+++ src/main/java/org/sopt/domain/post/presentation/message/PostMessage.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.post.controller.message;
+package org.sopt.domain.post.presentation.message;
 
 public class PostMessage {
     public static final String CREATED_SUCCESS = "성공적으로 게시글을 등록했습니다.";
