+++ src/main/java/org/sopt/domain/post/presentation/response/GetDetailedPostResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.post.controller.response;
+package org.sopt.domain.post.presentation.response;
 
 import org.sopt.domain.post.domain.Post;
 import org.sopt.domain.user.dto.response.GetUserResponse;
