+++ src/main/java/org/sopt/domain/post/presentation/response/SearchResultResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.post.controller.response;
+package org.sopt.domain.post.presentation.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
